Add patch to remove chain ID hex validation
The method `upsertNetworkController` has validation to ensure the chain
ID is a hex string. This was introduced here prematurely, before the
core packages were updated to use hex instead of decimal. This
validation has been removed with a patch, which we can then remove when
updating to the core v53 packages (which updates chain ID to be hex in
every package).